# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renewal Radar API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardMetrics {
  totalActive: number;
  overdue: number;
  dueSoon: number;
  completed: number;
  expired: number;
  paused: number;
  remindersSentLast30Days: number;
  byCategory: CategoryCount[];
}

export interface Workspace {
  id: number;
  name: string;
  slug: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWorkspaceBody {
  name: string;
  slug?: string;
}

export type WorkspaceMemberRole =
  (typeof WorkspaceMemberRole)[keyof typeof WorkspaceMemberRole];

export const WorkspaceMemberRole = {
  owner: "owner",
  admin: "admin",
  member: "member",
} as const;

export interface WorkspaceMember {
  id: number;
  workspaceId: number;
  clerkUserId: string;
  email: string;
  name?: string;
  role: WorkspaceMemberRole;
  joinedAt: string;
}

export type InviteMemberBodyRole =
  (typeof InviteMemberBodyRole)[keyof typeof InviteMemberBodyRole];

export const InviteMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface InviteMemberBody {
  email: string;
  role?: InviteMemberBodyRole;
}

export type ObligationStatus =
  (typeof ObligationStatus)[keyof typeof ObligationStatus];

export const ObligationStatus = {
  active: "active",
  expired: "expired",
  completed: "completed",
  paused: "paused",
} as const;

export type ObligationRenewalFrequency =
  | (typeof ObligationRenewalFrequency)[keyof typeof ObligationRenewalFrequency]
  | null;

export const ObligationRenewalFrequency = {
  once: "once",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
  custom: "custom",
} as const;

export interface Obligation {
  id: number;
  workspaceId: number;
  title: string;
  description?: string | null;
  category: string;
  status: ObligationStatus;
  dueDate: string;
  renewalFrequency?: ObligationRenewalFrequency;
  customFrequencyDays?: number | null;
  ownerClerkId?: string | null;
  ownerName?: string | null;
  ownerEmail?: string | null;
  backupOwnerClerkId?: string | null;
  backupOwnerName?: string | null;
  backupOwnerEmail?: string | null;
  notes?: string | null;
  tags: string[];
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateObligationBodyRenewalFrequency =
  | (typeof CreateObligationBodyRenewalFrequency)[keyof typeof CreateObligationBodyRenewalFrequency]
  | null;

export const CreateObligationBodyRenewalFrequency = {
  once: "once",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
  custom: "custom",
} as const;

export interface CreateObligationBody {
  workspaceId: number;
  title: string;
  description?: string | null;
  category: string;
  dueDate: string;
  renewalFrequency?: CreateObligationBodyRenewalFrequency;
  customFrequencyDays?: number | null;
  ownerClerkId?: string | null;
  ownerEmail?: string | null;
  backupOwnerClerkId?: string | null;
  backupOwnerEmail?: string | null;
  notes?: string | null;
  tags?: string[];
}

export interface CompleteObligationBody {
  notes?: string | null;
}

export type ReminderRuleChannel =
  (typeof ReminderRuleChannel)[keyof typeof ReminderRuleChannel];

export const ReminderRuleChannel = {
  email: "email",
  in_app: "in_app",
} as const;

export type ReminderRuleRecipientType =
  (typeof ReminderRuleRecipientType)[keyof typeof ReminderRuleRecipientType];

export const ReminderRuleRecipientType = {
  owner: "owner",
  backup_owner: "backup_owner",
  all_members: "all_members",
  custom_email: "custom_email",
} as const;

export interface ReminderRule {
  id: number;
  obligationId: number;
  daysBefore: number;
  channel: ReminderRuleChannel;
  recipientType: ReminderRuleRecipientType;
  customEmail?: string | null;
  isActive: boolean;
  lastTriggeredAt?: string | null;
  createdAt: string;
}

export type CreateReminderRuleBodyChannel =
  (typeof CreateReminderRuleBodyChannel)[keyof typeof CreateReminderRuleBodyChannel];

export const CreateReminderRuleBodyChannel = {
  email: "email",
  in_app: "in_app",
} as const;

export type CreateReminderRuleBodyRecipientType =
  (typeof CreateReminderRuleBodyRecipientType)[keyof typeof CreateReminderRuleBodyRecipientType];

export const CreateReminderRuleBodyRecipientType = {
  owner: "owner",
  backup_owner: "backup_owner",
  all_members: "all_members",
  custom_email: "custom_email",
} as const;

export interface CreateReminderRuleBody {
  /** @minimum 0 */
  daysBefore: number;
  channel: CreateReminderRuleBodyChannel;
  recipientType: CreateReminderRuleBodyRecipientType;
  customEmail?: string | null;
  isActive?: boolean;
}

export type DeliveryRecordStatus =
  (typeof DeliveryRecordStatus)[keyof typeof DeliveryRecordStatus];

export const DeliveryRecordStatus = {
  sent: "sent",
  failed: "failed",
  pending: "pending",
} as const;

export interface DeliveryRecord {
  id: number;
  obligationId: number;
  obligationTitle?: string | null;
  ruleId?: number | null;
  channel: string;
  recipientEmail: string;
  status: DeliveryRecordStatus;
  errorMessage?: string | null;
  sentAt: string;
}

export type AuditLogDetails = { [key: string]: unknown } | null;

export interface AuditLog {
  id: number;
  workspaceId?: number | null;
  obligationId?: number | null;
  obligationTitle?: string | null;
  actorClerkId?: string | null;
  actorName?: string | null;
  action: string;
  details?: AuditLogDetails;
  createdAt: string;
}

export type CsvPreviewBodyColumnMapping = { [key: string]: string };

export interface CsvPreviewBody {
  csvContent: string;
  columnMapping?: CsvPreviewBodyColumnMapping;
  workspaceId: number;
}

export type CsvPreviewResultRowsItem = { [key: string]: unknown };

export interface CsvPreviewResult {
  headers: string[];
  rows: CsvPreviewResultRowsItem[];
  totalRows: number;
  validRows: number;
  errors: string[];
}

export type CsvImportBodyColumnMapping = { [key: string]: string };

export interface CsvImportBody {
  csvContent: string;
  columnMapping: CsvImportBodyColumnMapping;
  workspaceId: number;
}

export interface CsvImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export type GetDashboardMetricsParams = {
  workspaceId?: number;
};

export type GetUpcomingObligationsParams = {
  workspaceId?: number;
  days?: number;
};

export type ListObligationsParams = {
  workspaceId?: number;
  status?: ListObligationsStatus;
  category?: string;
  search?: string;
  ownerId?: string;
  dueBefore?: string;
  dueAfter?: string;
};

export type ListObligationsStatus =
  (typeof ListObligationsStatus)[keyof typeof ListObligationsStatus];

export const ListObligationsStatus = {
  active: "active",
  expired: "expired",
  completed: "completed",
  paused: "paused",
} as const;

export type ListDeliveryHistoryParams = {
  workspaceId?: number;
  obligationId?: number;
  status?: ListDeliveryHistoryStatus;
  limit?: number;
};

export type ListDeliveryHistoryStatus =
  (typeof ListDeliveryHistoryStatus)[keyof typeof ListDeliveryHistoryStatus];

export const ListDeliveryHistoryStatus = {
  sent: "sent",
  failed: "failed",
  pending: "pending",
} as const;

export type ListAuditLogsParams = {
  workspaceId?: number;
  obligationId?: number;
  limit?: number;
};

export type ExportObligationsCsvParams = {
  workspaceId?: number;
  status?: string;
};