# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Renewal Radar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /obligations:
    get:
      operationId: listObligations
      tags: [obligations]
      summary: List obligations with optional filters
      security: [{ bearerAuth: [] }]
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, expired, completed, paused]
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: dueBefore
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dueAfter
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of obligations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obligation"
    post:
      operationId: createObligation
      tags: [obligations]
      summary: Create a new obligation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObligationBody"
      responses:
        "201":
          description: Created obligation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"
  /obligations/{obligationId}:
    get:
      operationId: getObligation
      tags: [obligations]
      summary: Get an obligation by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      responses:
        "200":
          description: Obligation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"
    put:
      operationId: updateObligation
      tags: [obligations]
      summary: Update an obligation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObligationBody"
      responses:
        "200":
          description: Updated obligation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"
    delete:
      operationId: deleteObligation
      tags: [obligations]
      summary: Delete an obligation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      responses:
        "204":
          description: Deleted
  /obligations/{obligationId}/complete:
    post:
      operationId: completeObligation
      tags: [obligations]
      summary: Mark an obligation as completed
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteObligationBody"
      responses:
        "200":
          description: Updated obligation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"

  # ─── Reminder Rules ──────────────────────────────────────────────────────────
...
  /obligations/export/csv:
    get:
      operationId: exportObligationsCsv
      tags: [csv]
      summary: Export obligations as CSV
      security: [{ bearerAuth: [] }]
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /obligations/import/csv/preview:
    post:
      operationId: previewCsvImport
      tags: [csv]
      summary: Preview CSV import with column mapping
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvPreviewBody"
      responses:
        "200":
          description: Preview of parsed rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvPreviewResult"
  /obligations/import/csv:
    post:
      operationId: importObligationsCsv
      tags: [csv]
      summary: Import obligations from CSV with column mapping
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvImportResult"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ObligationIdParam:
      name: obligationId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Obligation:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        status:
          type: string
          enum: [active, expired, completed, paused]
        dueDate:
          type: string
          format: date
        renewalFrequency:
          type: string
          enum: [once, monthly, quarterly, annually, custom]
          nullable: true
        customFrequencyDays:
          type: integer
          nullable: true
        ownerClerkId:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        backupOwnerClerkId:
          type: string
          nullable: true
        backupOwnerName:
          type: string
          nullable: true
        backupOwnerEmail:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - title
        - category
        - status
        - dueDate
        - tags
        - createdAt
        - updatedAt
    CreateObligationBody:
      type: object
      properties:
        workspaceId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        dueDate:
          type: string
          format: date
        renewalFrequency:
          type: string
          enum: [once, monthly, quarterly, annually, custom]
          nullable: true
        customFrequencyDays:
          type: integer
          nullable: true
        ownerClerkId:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        backupOwnerClerkId:
          type: string
          nullable: true
        backupOwnerEmail:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [workspaceId, title, category, dueDate]
    CompleteObligationBody:
      type: object
      properties:
        notes:
          type: string
          nullable: true

    # ── Reminder Rules ─────────────────────────────────────────────────────────
...
    CsvPreviewBody:
      type: object
      properties:
        csvContent:
          type: string
        columnMapping:
          type: object
          additionalProperties:
            type: string
        workspaceId:
          type: integer
      required: [csvContent, workspaceId]
    CsvPreviewResult:
      type: object
      properties:
        headers:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        totalRows:
          type: integer
        validRows:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [headers, rows, totalRows, validRows, errors]
    CsvImportBody:
      type: object
      properties:
        csvContent:
          type: string
        columnMapping:
          type: object
          additionalProperties:
            type: string
        workspaceId:
          type: integer
      required: [csvContent, columnMapping, workspaceId]
    CsvImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [imported, skipped, errors]