# lib/api-zod/src/generated/types/csvImportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renewal Radar API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CsvImportBodyColumnMapping } from "./csvImportBodyColumnMapping";

export interface CsvImportBody {
  csvContent: string;
  columnMapping: CsvImportBodyColumnMapping;
  workspaceId: number;
}